+++ Bot/Cogs/events_log.py
@@ -1,9 +1,8 @@
-from attrs import asdict
 from discord import PartialEmoji
 from discord.ext import commands
 from kumikocore import KumikoCore
 from Libs.cache import KumikoCache
-from Libs.cog_utils.events_log import EventsFlag, get_or_fetch_config
+from Libs.cog_utils.events_log import EventsFlag, get_or_fetch_channel_id
 from Libs.config import LoggingGuildConfig, get_or_fetch_guild_config
 from Libs.ui.events_log import RegisterView, UnregisterView
 from Libs.utils import ConfirmEmbed, Embed, is_manager
@@ -16,6 +15,7 @@ class EventsLog(commands.Cog):
         self.bot = bot
         self.pool = self.bot.pool
         self.redis_pool = self.bot.redis_pool
+        self.events_name_list = ["member_events", "mod_events", "eco_events"]
 
     @property
     def display_emoji(self) -> PartialEmoji:
@@ -77,9 +77,16 @@ class EventsLog(commands.Cog):
 
     @is_manager()
     @commands.guild_only()
-    @logs.command(name="configure", aliases=["config"])
-    async def config(self, ctx: commands.Context, events: EventsFlag) -> None:
-        """Configures which events are enabled"""
+    @logs.command(name="configure", aliases=["config"], usage="all: bool")
+    async def config(
+        self, ctx: commands.Context, name: str, status: bool, *, events: EventsFlag
+    ) -> None:
+        """Configures which events are enabled. Using the all flag enabled all events."""
+        if name not in self.events_name_list:
+            await ctx.send(
+                "The name of the event was not found. The possible events are:\nmember_events\nmod_events\neco_events"
+            )
+            return
         query = """
         UPDATE logging_config
         SET member_events = $2, mod_events = $3, eco_events = $4
@@ -88,30 +95,44 @@ class EventsLog(commands.Cog):
         guild_id = ctx.guild.id  # type: ignore
         key = f"cache:kumiko:{guild_id}:guild_config"
         cache = KumikoCache(connection_pool=self.redis_pool)
-        get_config = await get_or_fetch_config(
-            id=guild_id, redis_pool=self.redis_pool, pool=self.pool
+        get_channel_id = await get_or_fetch_channel_id(
+            guild_id=guild_id, pool=self.pool, redis_pool=self.redis_pool
         )
-        if get_config is None:
+        if get_channel_id is None:
             await ctx.send("The config was not set up. Please enable the logs module")
             return
 
+        statuses = {
+            "member_events": status if name in "member_events" else False,
+            "mod_events": status if name in "mod_events" else False,
+            "eco_events": status if name in "eco_events" else False,
+        }
+
         lgc = LoggingGuildConfig(
-            channel_id=get_config["channel_id"],
-            member_events=events.member,
-            mod_events=events.mod,
-            eco_events=events.eco,
+            channel_id=int(get_channel_id),
+            member_events=statuses["member_events"],
+            mod_events=statuses["mod_events"],
+            eco_events=statuses["eco_events"],
         )
         if events.all is True:
             lgc = LoggingGuildConfig(
-                channel_id=get_config["channel_id"],
+                channel_id=int(get_channel_id),
                 member_events=True,
                 mod_events=True,
                 eco_events=True,
             )
+            await self.pool.execute(query, guild_id, True, True, True)
+        else:
+            await self.pool.execute(
+                query,
+                guild_id,
+                statuses["member_events"],
+                statuses["mod_events"],
+                statuses["eco_events"],
+            )
 
-        await self.pool.execute(query, guild_id, events.member, events.mod, events.eco)
         await cache.merge_json_cache(
-            key=key, value=asdict(lgc), path="$.logging_config"
+            key=key, value=lgc, path=".logging_config", ttl=None
         )
         await ctx.send("Updated successfully!")