+++ Bot/Libs/cog_utils/events_log/cache_utils.py
@@ -20,7 +20,7 @@ async def get_or_fetch_config(
     key = f"cache:kumiko:{id}:guild_config"
     cache = KumikoCache(redis_pool)
     if await cache.cache_exists(key=key):
-        res = await cache.get_json_cache(key=key, path="$.logging_config")
+        res = await cache.get_json_cache(key=key, path=".logging_config")
         return res
     else:
         rows = await pool.fetchrow(query, id)
@@ -40,12 +40,33 @@ async def get_or_fetch_log_enabled(
     key = f"cache:kumiko:{id}:guild_config"
     cache = KumikoCache(redis_pool)
     if await cache.cache_exists(key=key):
-        res = await cache.get_json_cache(key=key, path="$.logs")
+        res = await cache.get_json_cache(key=key, path=".logs", value_only=False)
         return res  # type: ignore
     else:
         val = await pool.fetchval(query, id)
         if val is None:
             return False
+        return val
+
+
+async def get_or_fetch_channel_id(
+    guild_id: int, pool: asyncpg.Pool, redis_pool: ConnectionPool
+) -> Union[str, None]:
+    query = """
+    SELECT logging_config.channel_id
+    FROM logging_config
+    WHERE guild_id = $1;
+    """
+    key = f"cache:kumiko:{guild_id}:logging_channel_id"
+    cache = KumikoCache(redis_pool)
+    if await cache.cache_exists(key=key):
+        res = await cache.get_basic_cache(key=key)
+        return res
+    else:
+        val = await pool.fetchval(query, guild_id)
+        if val is None:
+            return None
+        await cache.set_basic_cache(key=key, value=val, ttl=3600)
         return val
 
 
@@ -61,14 +82,11 @@ async def set_or_update_cache(
         )
 
 
-async def delete_cache(key: str, redis_pool: ConnectionPool) -> None:
-    cache = KumikoCache(connection_pool=redis_pool)
-    if await cache.cache_exists(key=key):
-        await cache.delete_json_cache(key=key)
-
-
 async def disable_logging(guild_id: int, redis_pool: ConnectionPool) -> None:
     key = f"cache:kumiko:{guild_id}:guild_config"
     cache = KumikoCache(connection_pool=redis_pool)
-    await cache.merge_json_cache(key=key, value=False, path="$.logs")
-    await cache.merge_json_cache(key=key, value=None, path="$.logging_config")
+    await cache.delete_basic_cache(key=f"cache:kumiko:{guild_id}:logging_channel_id")
+    await cache.merge_json_cache(key=key, value=False, path=".logs", ttl=None)
+    await cache.merge_json_cache(
+        key=key, value=None, path=".logging_config.channel_id", ttl=None
+    )