+++ changelog.md
@@ -30,6 +30,9 @@ For the full list of changes, please see them here: [`v0.10.2...v0.11.0`](https:
 - Use grouped unique keys to enforce M-M relations
 - Update docs to reflect PEP8 standards
 - Improved traceback formatting for backwards compatibility
+- Use `msgspec.Struct` instead of `attrs` for faster serializations
+- Improve Redis caching
+- Improve event logs by using better redis caching and structs instead
 
 ## ✨ Additions